Revert "Create CNAME" to preserve fossrit.github.io domain
This reverts commit 356505733afae1720988b73aea19f2bc7f60cb64.

This decision is made to simplify the hosting for the FOSS@MAGIC
website. In the past, we used the fossr.it domain but lost control of
the domain. Unless the domain is put under control of MAGIC / SJ to
maintain, I prefer to keep the GitHub Pages domain as the default so
whoever is working on this in four or five years won't have to figure
out how the domain is set up or where it is hosted.