FIO: Add "Tracking the impact of zoo conservation" from Tom Snyder
This commit adds a new surprise speaker we got near the last minute.
This adds his name to the Speaker page as well as a new row for a Track
D. It doesn't look pretty as it is now, but maybe @ct-martin can take a
look and see if there is a better way of laying out the fourth track in
the schedule.

Signed-off-by: Justin W. Flory <git@jwf.io>