:building_construction: Modernize dev environment script (closes #36)
This commit drops the old `start_devel_env.sh` script in favor of a more
versatile `build.sh` script. This takes inspiration from tooling written
for the Fedora Docs project in using different container runtimes based
on what the user has installed to build a site (e.g. Docker or Podman).

I considered using Docker Compose for this, but I think the future of
rootless containers makes Docker Compose less interesting of a tool
choice, because it is locked into the Docker container runtime. This
script still uses Docker on macOS and Linux, but it also provides the
possibility to use other container runtimes like Podman on Linux. Thus,
I think this reworking of the dev environment build script is enough to
close out issue #36.

Note this commit does not update the docs, but this content will be
moved to the Runbook instead of this repo. This is being tracked in
FOSSRIT/fossrit.github.io#86.

Closes #36.

Signed-off-by: Justin W. Flory <git@jwf.io>