:memo: docs: Link out to FOSSRIT Runbook (closes #86)
This commit follows up on previous Runbook PRs, FOSSRIT/runbook#30 and
FOSSRIT/runbook#32. Now that contributing guidelines and advice about
maintaining this site lives in the Runbook, the duplicated info is
removed from this repo. Instead, the reader is redirected to the
Runbook.

Stay D.R.Y. - Don't Repeat Yourself!

Closes #86.

Signed-off-by: Justin W. Flory <git@jwf.io>