:gem: Add "White House Champions of Change" event, from @decause
This was an accidental find, but it was too good to not include on our
new website. I stumbled across the only running archive I know of for
the old FOSSbox website:

    http://www.samlucidi.com/fossbox/articles/decause/magic-centers-fossrit-program-sends-students-to-white-house-champions-of-change-event.html

In a best effort to preserve that data, I've moved it over into our new
site and decided to publish it under the _Events_ page.

I only wish I had pictures to include here.

cc: @decause @jenneh @ramstush @DAWacker @itprofjacobs @mansam

Signed-off-by: Justin W. Flory <git@jwf.io>