:bridge_at_night: projects: Update TeleIRC for 2020
This commit modernizes the project page for TeleIRC and brings it into
the current year. I also added some of our new contributors to the team
member list: @kennedy @10eMyrT @Zedjones.

Signed-off-by: Justin W. Flory <git@jwf.io>

Add TeleIRC logo to project profile per @Tjzabel's suggestion in #111

Instead of adding a copy of the logo to this repository, I decided to
embed the one from the RITlug website onto our site. While there is some
risk that the link could break someday, I hope that if the logo is ever
changed or updated, it can be updated in fewer places and be updated
with less work.

Nice idea on including the logo, @Tjzabel. :+1:

Signed-off-by: Justin W. Flory <git@jwf.io>