:hammer_and_wrench: Add Travis CI gated deployments (closes #44)
This commits adds Travis CI support for FOSSRIT/fossrit.github.io. It
builds the site with Jekyll, checks for HTML errors and broken links
with `html-proofer`, and if successful and run on `master`, it deploys
the latest build of the site from Travis.

I used the official Jekyll docs to set up testing, although there were
several things missing that took me a long time to figure out how to get
that part working:

    https://jekyllrb.com/docs/continuous-integration/travis-ci/

Never forget, `bundler exec`. :grimacing:

It is worth noting a few things are disabled in `html-proofer` that
realistically should be enabled. We _should_ check for alt tags, we
should _not_ have to ignore `/projects/` but we do because they are
appended with `.html` extensions for some reason, unlike the rest of the
site. The `permalink` metadata we set in project profiles is likely to
blame for this.

At time of commit/pull request, CI is failing. Another PR coming later
will fix the 404 errors and other broken links that it discovered while
putting this PR together.

Additionally, I used the official Travis CI docs for setting up
automatic deployments from Travis:

    https://docs.travis-ci.com/user/deployment/pages/

Additionally, I validated the Travis config in the beta Travis CI Config
explorer/validator tool:

    https://config.travis-ci.com/

Closes #44.

Signed-off-by: Justin W. Flory <git@jwf.io>

:memo: README: Add Travis CI testing badge

Show the latest status of `master` branch builds in Travis CI on the
README.

Signed-off-by: Justin W. Flory <git@jwf.io>