:link: about: Fix link to FOSS Minor curriculum
Similar to commit 3710bcf216d77703bdc46346e119c51e309c595d. The old link
to the PDF was dropped on RIT's website, this updates the pointer to the
current web home of the FOSS Minor.

Also, I added a small note for the Future is Open conference too on our
_About_ page.

Signed-off-by: Justin W. Flory <git@jwf.io>