:alembic: ci: Run tests with Rakefile, cache html-proofer data
This refactors the way the test suite is ran according to the
html-proofer documentation:

    https://github.com/gjtorikian/html-proofer/wiki/Using-HTMLProofer-From-Ruby-and-Travis

I like this approach better because it uses the existing "test" imports
in the `Gemfile` and all `html-proofer` config options are set in the
`Rakefile` instead of a really long command string in the `.travis.yml`.

Additionally, this also adds a check for valid HTML to `html-proofer`
since it is disabled by default. Seems like a useful thing to have.

Signed-off-by: Justin W. Flory <git@jwf.io>

:gem: Add rake to Gemfile, update Gemfile.lock

Signed-off-by: Justin W. Flory <git@jwf.io>

ci: Enable multiple caching features

https://docs.travis-ci.com/user/caching/#enabling-multiple-caching-features

Signed-off-by: Justin W. Flory <git@jwf.io>

ci: Use Ubuntu Bionic for build environment

Most notable benefit here is that Ruby 2.6.5 is already installed by
default on this image, so it doesn't need to be downloaded at the
beginning of the build like it does on Xenial. This should speed up CI
tests by a little.

Signed-off-by: Justin W. Flory <git@jwf.io>