:classical_building: projects: Add CIVX.us project profile
Last but not least, the CIVX.us project profile. This is the final
project I plan to archive in our new website before I depart RIT. I also
found an old RIT News article and decided to borrow the picture to embed
in our post. The news article is linked in the project profile.

Shout-outs and hat tips on this one go to @decause, @lmacken, and
@Qalthos. :tada:

Closes FOSSRIT/tasks#113.

Signed-off-by: Justin W. Flory <git@jwf.io>