:nut_and_bolt: ci: Revert Travis deploy pipeline
In favor of simplicity, this PR undoes the Travis CI deployment
pipeline. While all changes will _still_ be tested and required to pass
before new changes can be merged, Travis is no longer responsible for
pushing the built site source to the GitHub Pages environment.

The purpose of this change is to drop the `src` branch, revert `master`
to the default branch, and use a single branch for building and
deploying the site source. Since we are not using any advanced
functionality of Jekyll yet that warrants us to build outside of the
github-pages Gem, I think it is better to keep this simple for those who
will come after me.

Since this will bring us back to a fairly simple, single-branch workflow
with `master`, I no longer see the need in documenting the Runbook CI
deployment pipeline because there is no longer a pipeline. It goes back
to being a simple CI test.

Closes FOSSRIT/runbook#47.

Signed-off-by: Justin W. Flory <git@jwf.io>