:1234: projects: Add Blocku project profile (#133)
This commit adds another historical HFOSS project profile to the
FOSS@RIT website. This time, it covers Blocku, a Sudoku-style puzzle
game. This is also the first project profile to feature a moving GIF.
Neat! :camera_flash:

Shout-outs and hat tips on this one go to @ralphbean, @franrogers,
@eloreyen, @coolestdude1, @kaeedo2, @spectralshadow514, and
@Destroyer675000. :tada:

This was a helpful resource in compiling this project profile:

    https://github.com/ralphbean/tos-rit-projects-seminar/blob/master/data/students.yaml

Signed-off-by: Justin W. Flory <git@jwf.io>