update virtual foss hours url (*see also note)
as per @jwflofy's request

Note: while it seems that it might _technical_ work, I can not comfortably say that putting a normal URL in the `location` field is a good idea nor best practice. As far as I can tell, this behavior seems to work on at least some calendar clients, but is not definitively to-spec.

Based on what I can find online, the best practice would be to put something to the effect of "Jitsi/Online" in the `location` field and the URL in the `url` field (populated by the `redirect` front matter) as well as in the description/summary (since not all clients support the `url` field, including Google).