:world_map: config: Generate a sitemap and highlight it in robots.txt (#142)
This is a Search Engine Optimization (S.E.O.) tweak. A sitemap is a file
for search engines that creates a digital map of all the content
published on a website. Using a plugin, Jekyll will create one for
fossrit.github.io and put it at the root of the domain. Using a
`robots.txt` is a more explicit way we tell other search engines that a
sitemap is present and to go and use it for indexing.

Signed-off-by: Justin W. Flory <git@jwf.io>