:mag: seo-meta: Add meta tag for Google Search Console verification
This commit adds a new `<meta>` tag that appears in the `<head>` of
every page. It verifies that admin access to `fossrit.github.io` is
correctly given in the Google Search Console interface.

My intention is to use this tool to submit the sitemap to Google and
also promote specific pages on our site in Google.

Signed-off-by: Justin W. Flory <git@jwf.io>