:family: Add FOSSRIT People/Profiles link to site navbar (#148)
This adds an outlink to the FOSSRIT People page. This better integrates
the work done by @kennedy for the FOSSRIT Profiles page into our web
presence. It does not close FOSSRIT/people#72 though. We need to link
back to this website from the People page in order to close that issue.

Signed-off-by: Justin W. Flory <git@jwf.io>