Update Open @ RIT Page and FOSS hours times (#190)
* Point Open @ RIT Page

Fix for Issue #188

* History Re-added

Added history back to open-rit.md and tried to fix the mailing list email, which I don't believe the mailing list or email still exist...

Added how to contribute to this repository and what website this repository is for in the README.

* Undid changes to email

Had to undo changes to email since it did not pass circleci build...

* Temporary History File

Made a temporary history file to keep track of the history of Open Source Groups at RIT before the page is made.

* Runbook, FOSS Hours Fix

RE-added runbook link in case there is useful information available there. Fixed FOSS Hours for this semester...
Also added a section to give credit to those who contributed information for the history of Open Source @ RIT.

* Requested Changes

Made the changes to the how to contribute, their new website, and their summary page.